requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
openai==0.28.1
pytz==2023.3